refactor: pacing, major abstraction, performance optimization (#8)
* feat: pace block propogation to frontend

* refactor: huge performance optimization, fix possible memory leak

* refactor: further performance boosts

* refactor: legacy & v0 handling abstraction

* feat: in-app queue for pacing

* refactor: modularize block processing, single-pass tx sort

* refactor: abstraction, performance, error-checks

* refactor: legacy & v0 abstraction

* refactor: further abstract v0 and legacy
