+++ package.json
@@ -24,6 +24,7 @@
     "bs58": "^5.0.0",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
+    "fastq": "^1.15.0",
     "superstruct": "^1.0.3",
     "ws": "^8.13.0"
   }
